+++ book.rb
@@ -1,13 +1,13 @@
 class Book
-  attr_accessor :title, :author, :rental
+  attr_accessor :title, :author, :rentals
 
   def initialize(title, author)
     @title = title
     @author = author
-    @rental = []
+    @rentals = []
   end
 
-  def add_rental(person, date)
-    Rental.new(date, self, person)
+  def add_rental(rental)
+    @rentals.push(rental)
   end
 end
