+++ create_teacher.rb
@@ -1,0 +1,16 @@
+require './teacher'
+
+module CreateTeacher
+  def create_teacher
+    puts 'Create a new teacher'
+    print 'Enter teacher age: '
+    age = gets.chomp.to_i
+    print 'Enter teacher name: '
+    name = gets.chomp
+    print 'Enter teacher specialization: '
+    specialization = gets.chomp
+    teacher = Teacher.new(age, specialization, name)
+    @people << teacher
+    puts "Teacher #{name}, created successfully"
+  end
+end
