+++ display_rental.rb
@@ -1,0 +1,14 @@
+module DisplayRental
+  def list_rentals
+    puts 'Select person by ID: '
+    @people.each do |person|
+      puts "ID: #{person.id}, Name: #{person.name}, Age: #{person.age}"
+    end
+    id = gets.chomp.to_i
+
+    puts 'Rentals:'
+    @rentals.each do |rental|
+      puts "Date: #{rental.date}, Book: #{rental.book.title}' by #{rental.book.author}" if rental.person.id == id
+    end
+  end
+end
