+++ rental.rb
@@ -5,6 +5,7 @@ class Rental
     @date = date
     @book = book
     book.rentals << self
+
     @person = person
     person.rentals << self
   end
